requirements.txt
fastapi~=0.95.0
uvicorn==0.21.1
sqlalchemy~=2.0.7
psycopg2-binary
pydantic~=1.10.6
bcrypt
passlib[bcrypt]
python-jose[cryptography]
feedparser~=6.0.10
asyncio
httpx
pyjwt
python-dotenv~=1.0.0
pytest~=7.1.2
python-dateutil